2019: Provide iter::permutation polyfill
`std::iter` doesn't provide an iterator for permutations, and a core
component of my AoC journey is to not use third-party libraries.

The permutation algorithm is known as Heap's Algorithm. It optimizes for
least number of element swaps and is not guaranteed to be lexicographic.
Additionally, the straightforward iterative pseudocode on Wikipedia was
easy to adapt to Rust's iterator trait protocol.